przeniesienie plikow do odpowiednich folderow - zamiana starych wersji na nowe
