+++ src/pso.c
@@ -1,62 +1,89 @@
 #include "pso.h"
 #include <time.h>
 #include <stdlib.h>
+#include <stdio.h>
+
+pso_settings read_pso_settings(const char* file_path){
+    pso_settings ret =(pso_settings){1,1,1};
+    FILE* fptr = fopen(file_path, "r");
+    if (fptr == NULL) {
+            printf("Błąd przy czytaniu mapy z pliku!\n");
+            return (pso_settings){1,1,1};
+    }
+
+    fscanf(fptr,"%lf %lf %lf ", &ret.w, &ret.c1, &ret.c2);
+    fclose(fptr);
+    return ret;
+}
 
 swarm* swarm_construct(int particle_number, Map * map){
     srand(time(NULL));
-    swarm * s = malloc(sizeof(swarm*));
+    swarm * s = malloc(sizeof(swarm));
     s->size = particle_number;
-    s->lowest_found_x = 0;
-    s->lowest_found_y = 0;
+    s->highest_found_x = 0;
+    s->highest_found_y = 0;
     s->particle_arr = malloc(sizeof(particle*)*particle_number);
-
     for(int i =0; i<particle_number; i++){
-        s->particle_arr[i] = malloc(sizeof(particle));
-        s->particle_arr[i]->pos_x = rand()%(map->width);
-        s->particle_arr[i]->pos_x = rand()%(map->heigth);
-        s->particle_arr[i]->vel_x = 0;
-        s->particle_arr[i]->vel_y = 0;
-        s->particle_arr[i]->lowest_found_x = 0;
-        s->particle_arr[i]->lowest_found_y = 0;
+        particle * p = malloc(sizeof(particle));
+
+        p->pos_x = rand()%(map->width);
+        p->pos_y = rand()%(map->heigth);
+        p->vel_x = 0;
+        p->vel_y = 0;
+        p->highest_found_x = (int)(p->pos_x);
+        p->highest_found_y = (int)(p->pos_y);
+
+        int ind = (int)p->pos_x + (int)p->pos_y * map->width;
+        int highest_ind = s->highest_found_x + s->highest_found_y * map->width;
+        if(map->depth->value[ind] > map->depth->value[highest_ind]){
+            s->highest_found_x = (int)p->pos_x;
+            s->highest_found_y = (int)p->pos_y;
+        }
+
+        s->particle_arr[i] = p;
     }
     return s;
 }
 // TU JAKO ARGUMENTY TRZEBA DODAC WAGI I WPISAC JE W ROWNANIA
-void update_particles(swarm * s, Map *m){
+void update_particles(swarm * s, Map *m, pso_settings pso_s){
     //dla każdej cząstki
     for(int i =0; i<s->size; i++){
         particle * p =  s->particle_arr[i];
-        //zmieniamy pozycje o predkosc
-        p->pos_x += p->vel_x;
-        p->pos_y += p->vel_y;
-        int ind = (int)p->pos_x + (int)p->pos_y * m->width;
-        int lowest_ind = p->lowest_found_x + p->lowest_found_y * m->width;
-        //sprawdzamy czy nowa pozycja to jest jakieś minimum
-        // tutaj lokalne dla cząstki
-        if(m->depth->value[ind] < m->depth->value[lowest_ind]){
-            p->lowest_found_x = (int)p->pos_x;
-            p->lowest_found_y = (int)p->pos_y * m->width;
-        }
-        int lowest_ind = s->lowest_found_x + s->lowest_found_y * m->width;
-        //tutaj globalne
-        if(m->depth->value[ind] < m->depth->value[lowest_ind]){
-            s->lowest_found_x = (int)p->pos_x;
-            s->lowest_found_y = (int)p->pos_y * m->width;
-        }
         //losujemy prędkość - składowa x
         float r1 = (float)rand()/(float)(RAND_MAX);
         float r2 = (float)rand()/(float)(RAND_MAX);
-        p->vel_x = 1 * p->vel_x + r1 * 1 * (p->lowest_found_x - p->pos_x) + r2 * 1 * (s->lowest_found_x - p->pos_x);
+        p->vel_x = pso_s.w * p->vel_x + r1 * pso_s.c1 * (p->highest_found_x - p->pos_x) + r2 * pso_s.c2 * (s->highest_found_x - p->pos_x);
         //jesli spowoduje ona, ze czastka wychodzi za mape to zmieniamy kierunek na przeciwny
-        if(p->pos_x + p->vel_x > m->width){
+        if((p->pos_x + p->vel_x > m->width) || (p->pos_x + p->vel_x < 0)){
             p->vel_x *= -1;
         }
         r1 = (float)rand()/(float)(RAND_MAX);
         r2 = (float)rand()/(float)(RAND_MAX);
-        p->vel_y = 1*p->vel_y + r1 * 1 * (p->lowest_found_y - p->pos_y) + r2 * 1 * (s->lowest_found_y - p->pos_y);
+        p->vel_y = pso_s.w * p->vel_y + r1 * pso_s.c1 * (p->highest_found_y - p->pos_y) + r2 * pso_s.c2 * (s->highest_found_y - p->pos_y);
         //to samo dla skladowej y
-        if(p->pos_y + p->vel_y > m->heigth){
+        if((p->pos_y + p->vel_y > m->heigth) || (p->pos_y + p->vel_y < 0)){
             p->vel_y *= -1;
+        }
+
+        printf(" PREDKOSC %lf %lf \n", p->vel_x, p->vel_y);
+
+        //zmieniamy pozycje o predkosc
+        p->pos_x += p->vel_x;
+        p->pos_y += p->vel_y;
+
+                int ind = (int)p->pos_x + (int)p->pos_y * m->width;
+        int highest_ind = p->highest_found_x + p->highest_found_y * m->width;
+        //sprawdzamy czy nowa pozycja to jest jakieś minimum
+        // tutaj lokalne dla cząstki
+        if(m->depth->value[ind] > m->depth->value[highest_ind]){
+            p->highest_found_x = (int)p->pos_x;
+            p->highest_found_y = (int)p->pos_y;
+        }
+        highest_ind = s->highest_found_x + s->highest_found_y * m->width;
+        //tutaj globalne
+        if(m->depth->value[ind] > m->depth->value[highest_ind]){
+            s->highest_found_x = (int)p->pos_x;
+            s->highest_found_y = (int)p->pos_y;
         }
     }
 }
